Hashmap legit saves lives, implement soon, problems with continue loop

+++ Pizza/Pizza/src/main/java/com/pizza/prices/Sizes.java
@@ -9,7 +9,7 @@ public class Sizes {
     public static int largePie = 15;
 
     public static String [] sizeArr = {"Small", "Medium", "Large"};
-    public static int [] sizePrice = {smallPie, medPie, largePie};
+    public static double[] sizePrice = {smallPie, medPie, largePie};
 
         public Sizes(int smallPie, int medPie, int largePie) {
                 Sizes.smallPie = smallPie;
